create a centralized logger
